+++ Demo/src/androidTest/java/com/paypal/android/CardTest.kt
@@ -2,7 +2,7 @@ package com.paypal.android
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.paypal.android.testutils.AppDriver
-import org.junit.Assert.assertTrue
+import org.junit.Assert.assertEquals
 import org.junit.Test
 import org.junit.runner.RunWith
 
@@ -17,16 +17,24 @@ class CardTest {
             launchAppFromHomeScreen()
 
             findText("CARD").click()
-            waitForText("Card Number")
+            findText("New Visa").click()
 
-            findText("Card Number").text = "4111111111111111"
-            findText("Expiration").text = "02/2023"
-            findText("Security Code").text = "123"
+            waitForText("CARD NUMBER")
 
-            findText("SUBMIT").click()
+            findText("SCA").click()
+            findText("WHEN REQUIRED").click()
 
-            waitForText("CAPTURE success: CONFIRMED")
-            assertTrue(findText("CAPTURE success: CONFIRMED").exists())
+            findText("INTENT").click()
+            findText("CAPTURE").click()
+
+            findText("SHOULD VAULT").click()
+            findText("NO").click()
+
+            findText("CREATE & APPROVE ORDER").click()
+
+            waitForText("Status: COMPLETED")
+            val statusText = findResById("statusText")
+            assertEquals("Status: COMPLETED", statusText.text)
         }
     }
 }