+++ Demo/src/main/java/com/paypal/android/data/card/TestCard.kt
@@ -1,5 +1,14 @@
 package com.paypal.android.data.card
 
+import android.os.Parcelable
 import com.paypal.android.cardpayments.Card
+import com.paypal.android.ui.card.validation.CardFormatter
+import kotlinx.parcelize.IgnoredOnParcel
+import kotlinx.parcelize.Parcelize
 
-data class TestCard(val name: String, val card: Card)
+@Parcelize
+data class TestCard(val name: String, val card: Card) : Parcelable {
+
+    @IgnoredOnParcel
+    val formattedCardNumber: String = CardFormatter.formatCardNumber(card.number)
+}