+++ Demo/src/main/java/com/paypal/android/ui/StringResourceList.kt
@@ -1,0 +1,9 @@
+package com.paypal.android.ui
+
+import androidx.annotation.StringRes
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.res.stringResource
+
+@Composable
+fun stringResourceListOf(@StringRes vararg resources: Int): List<String> =
+    resources.map { stringResource(id = it) }