+++ Demo/src/main/java/com/paypal/android/ui/testcards/TestCardsItem.kt
@@ -1,8 +1,0 @@
-package com.paypal.android.ui.testcards
-
-import com.paypal.android.data.card.TestCard
-
-sealed class TestCardsItem {
-    class Header(val title: String) : TestCardsItem()
-    class Data(val testCard: TestCard) : TestCardsItem()
-}