+++ Demo/src/main/java/com/paypal/android/ui/testcards/TestCardsListener.kt
@@ -1,8 +1,0 @@
-package com.paypal.android.ui.testcards
-
-import com.paypal.android.data.card.TestCard
-
-interface TestCardsListener {
-
-    fun onTestCardSelected(testCard: TestCard)
-}